+++ src/background/feeder/web2-helper.feeder.ts
@@ -1,3 +1,4 @@
+import { isBeeApiAvailable } from '../../utils/bee-js'
 import { InterceptorReqMessageFormat, ResponseMessageFormat } from '../../utils/message/message-handler'
 import { BeeApiListener } from '../listener/bee-api.listener'
 
@@ -19,7 +20,29 @@ export class Web2HelperFeeder {
           answer: this.beeApiListener.beeApiUrl,
         }
         sendResponse(response)
+      } else if (message.key === 'isBeeApiAvailable') {
+        console.log('Web2HelperFeeder:beeApiAvailable got aimed message from content script', message)
+
+        this.checkConnection(message, sendResponse)
+
+        return true
       }
     })
+  }
+
+  private async checkConnection(
+    message: InterceptorReqMessageFormat<string>,
+    sendResponse: (response?: ResponseMessageFormat) => void,
+  ) {
+    const answer = await isBeeApiAvailable(this.beeApiListener.beeApiUrl)
+
+    const response: ResponseMessageFormat = {
+      key: message.key,
+      sender: 'background',
+      target: 'content',
+      answer,
+    }
+
+    sendResponse(response)
   }
 }